+++ CHANGELOG.md
@@ -5,7 +5,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-- _empty_
+### Breaking
+- new template string format
+- removed `getDimensionsCss()` --> use `getCss()`
+
+### Added
+- getCss( size )
+- size alias support
+- fill color support (handy for icon systems)
 
 ## [1.4.1] - 2018-06-29
 ### Fixed